+++ CHANGELOG.md
@@ -25,6 +25,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.4] - 2023-04-19
+
+### Changed
+- Modified the `csm_api_client.k8s.load_kube_api()` helper function to prioritize
+  loading the kube config from inside the cluster, and to fall back to the config
+  file if the in-cluster config can't be read.
+
 ## [1.1.3] - 2023-04-10
 
 ### Fixed